Added a buy upgrade button
